Just tried the following program:
#lang typed-scheme

(require/typed
  srfi/1
  (fold (All (a b) ((a b -> b) b (Listof a) -> b))))

And got an error because (orig-module-stx) was #f.  I'm not sure whether
it should have been or not, but I've fixed up this to handle that case
(having it just use the normal error message when that's #f, as I'm not
sure whether the error message would be correct otherwise).

However, it might just be that (orig-module-stx) should have not been
#f, in which case this should be checked out.