There is a problem with empty lines: having a line (= a row) in the code tables with empty contents makes the row not appear.  Usually, we'd put an &nbsp; there so it does show, but that would be bad for the preprocessor examples where I really want to have the table contents reflect the exact file contents.
So another solution is to use a 'newline element, but then latex barfs
because it's not happy with a \\ inside a tt macro.  The hacked
"solution" is to have the newline element not be inside a tt element
-- latex doesn't barf now, but it does have one extra newline as a
result since it doesn't hide the newline to begin with.

A better solution is to find a way to make table cells in html not be
hidden when they have no contents, maybe through some css magic.

svn: r14206