Change the drdr:command-line property.
It now holds a list that starts with a known symbol, and a `*' symbol
marks where the file name is supposed to be.  #f means don't run the
file.  The conversion happens when the string is converted to a value,
and at that point it will complain if ~s is missing, if there's more
than one ~s, or if the command is not one of a known few commands.