Add proxies for -> and ->*.
Attempt to do a recently intelligent job by using more specific lambdas
when possible (i.e., when the arity of the function matches the arity
specified by the contract).

Also, change testcases that test creation of proxy contracts to stop
depending on the proxy-ness of -> contracts.  This also uncovered a bug
in the testcase that checks that proxy contracts aren't allowed for
the domain in hash/c.