Rearrange ->d appropriately to use proxies/chaperones.
Also remove the old tests for tail-call behavior with ->d.
I think it was a faulty way of doing the optimization, and we can't even
do it in the same way now, since we can't get the original "raw" arguments
to the right place via continuation marks.

->i has a much better chance of having this optimization where possible,
and when I start converting ->i, I'll make sure to add appropriate tests.