Spit out a warning if there are no files rendered.
Fix a reference to meta/web to be symbolic -- otherwise things can get
confused when symlinks are used, ending up with two instances of a
single modeule (once for the symbolic reference and one for a relative
one).  It's still best to avoid symlinks, though.

Simplify version specs in the techreports code, re-enable continue
reference, some reformatting.