Change the linux platform names.
For example, instead of "Linux - Fedora 7 (x86_64)", use "Linux x86_64
(Fedora 7)".  Hopefully, this will make it clearer that it's often fine
to use some linux installer on a different platform than the one it was
made on.  (For example, some people wanted an AMD64 build for Ubuntu,
and eventually it turned out that the x86_64 build on Fedora worked
fine.)

This is even more relevant now since the C part is much less platform
dependent in the linux distro sense.