Poll mirrors for the download web pages.
This is triggered by an environment variable that points at a file that
is used to cache polling results.  If the variable is not specified just
use all mirror links as usual.  This way, for random builds when people
try things out there is no problem, while the nightly builds (and my
manual builds when there are changes) do the right thing.

The file holds information about the polls, and verifies that the
download exists -- if not, the link is not shown.  If it is, then it
tries to get the size of the remote file (via HEAD or via FTP), and if
it doesn't match our download information, then drop it too.  If no size
informaion is available, include it, but re-poll after a few days.  Even
if the size matches, a re-poll will be done after a month, so stale
mirrors will not stick.