path normalization in setup/scribble
Needed so that `raco setup' doesn't rebuild documents as it
decides, say, that the "games/cards" document depends on
"games/scribblings\\games.scrbl" instead of
"games/scribblings/games.scrbl".

Has this been needed for a while, or did it somehow
become needed with recent changes to `raco setup' to
support `raco link'?