Move the dmg build to weatherwax (which can now deal with it. )
Also, remove attempt for a smart use of `fmt' -- the problem is that now
"$platform" is not set, so the conditional didn't do anything.  It would
be easy to get a conditional using `uname', but better to drop the whole
thing.
(cherry picked from commit 2a1464b49ee06a85a269dfc94ffb8abe590858f3)