Include the platform in the windows installer's "human name".
(Otherwise having both installed can be confusing since the uninstaller
has the same name; also, it's good to make a reminder for people who
want the other one.)

Also, just use the platform string in the registry key uniformly.

Also add a comment about detecting Win64 if it's desirable in the
future.