tighten bound for Mac OS X 10.6 localtime() hack
It looks like my bound for last time was too conservative,
in that I looked for the lowest number that didn't seem
to fail in 10.6. The range of failing values is apparently
not continuous.

I've tightened the bound to match the lowest
number that produces a useful result on my 10.7 machine,
assuming that it works for a continuous range there.
(The new bound is higher than the number previously used as
a lower bound.)

Merge to 5.2.1
(cherry picked from commit ef846caaa14c185ffe4e46a1a4f119a35ff37c34)