Add rel="nofollow" to short example links.
The examples make up the majority of the front page, and it turns out
that google used some of these links in "sitelinks" in search result.
These links should be the main entry points, and we ended up having such
gems as "Printf" be there.  They provide a way to "demote" URLS so they
won't show there, but doing that for one link means that another will
pop in (and there's a limit of a 100 such demotions).  So add this in an
attempt to make it not consider them as things that can be used in the
sitelinks.  This might be reverted if it doesn't help.

(Note that this is discouraged as a way to "sculpt pagerank", but the
purpose here is not to affect how they crawl the web pages and make page
reputation flow, but only to cover up for a missing feature that ends up
with horrible sitelinks.  The pages are scanned anyway from the doc
pages.)