Changing contracts on gl-context<%> and applying it to the gl-context% object.
I updated and noticed that this was now failing because interface
contracts are not check structurally (any more?). Once I applied the
interface, the contract application failed because the documented two
additional arguments (alternative and enable-breaks?) are not actually
accepted. I look through all the implementations of gl-context% and
none of them accept these. Should I change the documentation or is
this a bad backwards incompatibility with the older versions?