revert use of `lazy-require' in `racket/match' implementation
Using `lazy-require' under `begin-for-syntax' expands to a use of
`define-runtime-path' under `begin'for-syntax'. Unfortunately,
`define-runtime-path' doesn't yet work with `raco exe' when it
appears under `begin-for-syntax'. Although `define-runtime-path'
should be fixed, it may take a while. Meanwhile, reverting the
change allows programs that require `racket/gui/base' to work
with `raco exe'.

There's another `lazy-require' under `begin-for-syntax' that
doesn't seem to cause the same trouble, though.