Added `parallel-array->mutable-array' and `parallel-array-strict'
It's currently quite difficult to coax extra performance out of these.
The array has to be huge and the operations have to either not allocate
or allocate very little. Also, the computation is divided up very
naively. Still, the framework and main ideas are here, and James can
monkey with it all he likes.