`raco pkg': treat ".plt" format consistently with `raco setup -A'
An old package intended for use with `raco setup -A' can be
installed with `raco pkg install'.

A package created with `raco pkg create --format plt' can be
installed with `raco setup -A', although it could leave behind
weird "MANIFEST.rktd" and other files in the "collects" directory.

An old package created with `raco pkg create --format plt' still
can be used with `raco pkg install', although not with `raco setup -A'.

The change from previous 'raco pkg' behavior is that package content
now claims to be in a "collects" directory that is
installation-relative, and unpacking redirects "collects" to the
package-staging area. At the same time, unpacking still works the
same as before on ".plt" archives that are not installation-relative.