Add escape of parens in href strings, to make the use of such URLs more robust with external tools.
External tools on the web often incorrectly delimit the
URLs used in Scribble anchors when it contains parens.  Mis-delimiting
the URL causes the link to break.  Jeff Atwood recommends escaping
parens as well that tools such as Wikipedia and Stack Overflow don't
fail so badly:  http://www.codinghorror.com/blog/2008/10/the-problem-with-urls.html