Do the conventional DMG layout.
Note that this requires a running Finder.  Note also that the process
involves attaching the created DMG which means that it might be
problematic with a machine that might have a volume by the same name
already attached (since the applescript uses the volume name, which is
not required to be unique).
(cherry picked from commit b5618b7d139d1d30d2a82a681255bc185cb6dcc7)